BUTTONS :: update 2021-10-14-1
* Removed redundant functions
* removed auto theming, in facor of the color behavior management and 
events.
* re worked base buttons and removed parameters that are now shared 
inside the color behavior class.
* _text_color is the new front end property that will manage color in 
the MDIcon rather than the old direct _text_color.  this allows us to 
simply change the backend color and turn it back to it's "default state" 
that is the front end text_color.
* In favor of the new update on kivy[master], the line buttons now parse 
the radius parameter directly to the Line: instructions.
* Lines now uses _line_color for the actual color of the vertex 
instruction.
* removed fill roundFlatIconButton code in favor of the new behaviors 
inherited.
* the Segments property is now applied to line instructions and to the 
canvas instructions whenever we do a ripple animation.